+++ net/cpp/Socket.h
@@ -176,7 +176,7 @@ public:
     * 
     * @param timeout the send timeout in milliseconds.
     */
-   virtual void setSendTimeout(unsigned long long timeout) = 0;
+   virtual void setSendTimeout(unsigned long timeout) = 0;
    
    /**
     * Gets the send timeout for this Socket. This is the amount of time that
@@ -184,7 +184,7 @@ public:
     * 
     * @return the send timeout in milliseconds.
     */
-   virtual unsigned long long getSendTimeout() = 0;
+   virtual unsigned long getSendTimeout() = 0;
    
    /**
     * Sets the receive timeout for this Socket. This is the amount of time that
@@ -192,7 +192,7 @@ public:
     * 
     * @param timeout the receive timeout in milliseconds.
     */
-   virtual void setReceiveTimeout(unsigned long long timeout) = 0;
+   virtual void setReceiveTimeout(unsigned long timeout) = 0;
    
    /**
     * Gets the receive timeout for this Socket. This is the amount of time that
@@ -200,7 +200,7 @@ public:
     * 
     * @return the receive timeout in milliseconds.
     */
-   virtual unsigned long long getReceiveTimeout() = 0;
+   virtual unsigned long getReceiveTimeout() = 0;
    
    /**
     * Gets the number of Socket connections that can be kept backlogged while
