+++ net/cpp/SocketWrapper.h
@@ -202,7 +202,7 @@ public:
     * 
     * @param timeout the send timeout in milliseconds.
     */
-   virtual void setSendTimeout(unsigned long long timeout);
+   virtual void setSendTimeout(unsigned long timeout);
    
    /**
     * Gets the send timeout for this Socket. This is the amount of time that
@@ -210,7 +210,7 @@ public:
     * 
     * @return the send timeout in milliseconds.
     */
-   virtual unsigned long long getSendTimeout();
+   virtual unsigned long getSendTimeout();
    
    /**
     * Sets the receive timeout for this Socket. This is the amount of time that
@@ -218,7 +218,7 @@ public:
     * 
     * @param timeout the receive timeout in milliseconds.
     */
-   virtual void setReceiveTimeout(unsigned long long timeout);
+   virtual void setReceiveTimeout(unsigned long timeout);
    
    /**
     * Gets the receive timeout for this Socket. This is the amount of time that
@@ -226,7 +226,7 @@ public:
     * 
     * @return the receive timeout in milliseconds.
     */
-   virtual unsigned long long getReceiveTimeout();
+   virtual unsigned long getReceiveTimeout();
    
    /**
     * Gets the number of Socket connections that can be kept backlogged while
